refactor: 💡 add  SlidingPanel component and update Cryptopunks
