+++ src/components/DataRoom/CryptoPunks/styles.module.css
@@ -11,19 +11,7 @@
   display: flex;
 }
 
-.rightPanelContainer {
-  width: 50%;
-  height: 100%;
-  position: absolute;
-  right: 0;
-  color: var(--mui-palette-text-dark);
-  overflow-x: hidden;
-  display: flex;
-  align-items: flex-end;
-  padding-top: 64px;
-}
-
-.rightPanelContent {
+.slidingPanelContent {
   z-index: 20;
   display: flex;
   flex-direction: column;
@@ -31,14 +19,6 @@
   padding: 64px;
   height: 100%;
   gap: 30px;
-}
-
-.rightPanelBG {
-  position: absolute;
-  inset: 0;
-  background-color: var(--mui-palette-primary-main);
-  z-index: 10;
-  margin-top: 72px;
 }
 
 .percentage {
@@ -65,7 +45,7 @@
   align-items: baseline;
 }
 
-.leftPanelContainer {
+.punksGrid {
   width: 100%;
   height: 100%;
   overflow: hidden;
@@ -94,37 +74,23 @@
     flex-direction: column-reverse;
   }
 
-  .leftPanelContainer {
+  .punksGrid {
     padding-top: 72px;
     width: 100%;
     height: 100%;
+    overflow: hidden;
   }
 
-  .rightPanelContainer {
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    bottom: 0;
-  }
-
-  .rightPanelContent {
+  .slidingPanelContent {
     justify-content: end;
     padding: 30px;
     gap: 30px;
-  }
-
-  .rightPanelBG {
-    margin-top: 64px;
   }
 
   .statsContainer {
     display: flex;
     flex-direction: column;
     gap: 30px;
-  }
-
-  .leftPanelContainer {
-    overflow: hidden;
   }
 
   .cryptopunk {
