+++ src/components/DataRoom/IndustryComparison/index.tsx
@@ -1,12 +1,13 @@
 import type { BaseBlock } from '@/components/Home/types'
-import type { MotionValue } from 'framer-motion'
 import { useTransform, motion, useScroll } from 'framer-motion'
-import type { ReactNode } from 'react'
+import dynamic from 'next/dynamic'
 import React, { useRef } from 'react'
 import css from './styles.module.css'
 import { useIsMediumScreen } from '@/hooks/useMaxWidth'
 import { Typography } from '@mui/material'
 import DotGrid from './DotGrid'
+
+const SlidingPanel = dynamic(() => import('@/components/common/SlidingPanel'))
 
 const IndustryComparison = ({ title }: BaseBlock) => {
   const backgroundRef = useRef<HTMLDivElement>(null)
@@ -18,52 +19,30 @@ const IndustryComparison = ({ title }: BaseBlock) => {
     offset: ['start end', 'end start'],
   })
 
+  const opacityParams = [0.25, 0.35, 0.65, 0.75]
+  const opacity = useTransform(scrollYProgress, opacityParams, [0, 1, 1, 0])
+
   return (
     <div ref={backgroundRef} className={css.sectionContainer}>
-      <div className={css.stickyContainer}>
-        <RightPanel containerRef={gridContainerRef} scrollYProgress={scrollYProgress} isMobile={isMobile}>
-          <Typography className={css.title} variant="h1">
-            {title}
-          </Typography>
-          <DotGrid containerRef={gridContainerRef} scrollYProgress={scrollYProgress} />
-        </RightPanel>
+      <div ref={gridContainerRef} className={css.stickyContainer}>
+        <SlidingPanel
+          scrollParams={isMobile ? [0, 1] : [0.25, 0.35, 0.65, 0.75]}
+          translateParams={isMobile ? ['0%', '0%'] : ['100%', '0%', '0%', '100%']}
+          scrollYProgress={scrollYProgress}
+        >
+          <motion.div
+            className={css.slidingPanelContent}
+            style={{
+              opacity: isMobile ? 1 : opacity,
+            }}
+          >
+            <Typography className={css.title} variant="h1">
+              {title}
+            </Typography>
+            <DotGrid containerRef={gridContainerRef} scrollYProgress={scrollYProgress} />
+          </motion.div>
+        </SlidingPanel>
       </div>
-    </div>
-  )
-}
-
-const RightPanel = ({
-  scrollYProgress,
-  children,
-  containerRef,
-  isMobile,
-}: {
-  scrollYProgress: MotionValue<number>
-  children: ReactNode
-  isMobile: boolean
-  containerRef: React.RefObject<HTMLDivElement>
-}) => {
-  const opacityParams = [0.25, 0.35, 0.65, 0.75]
-  const translateParams = [0.25, 0.35, 0.65, 0.75]
-  const opacity = useTransform(scrollYProgress, opacityParams, [0, 1, 1, 0])
-  const bgTranslate = useTransform(scrollYProgress, translateParams, ['100%', '0%', '0%', '100%'])
-
-  return (
-    <div ref={containerRef} className={css.rightPanelContainer}>
-      <motion.div
-        className={css.rightPanelContent}
-        style={{
-          opacity: isMobile ? 1 : opacity,
-        }}
-      >
-        {children}
-      </motion.div>
-      <motion.div
-        className={css.rightPanelBG}
-        style={{
-          translateX: isMobile ? '0%' : bgTranslate,
-        }}
-      ></motion.div>
     </div>
   )
 }
